Add typeorm and setup db connection
