+++ README.md
@@ -62,7 +62,3 @@ npm ci
 
 - web development framework for Node.js
 - server-side/backend framework which is used for API or rendering apps
-
-# Database - postgres/mysql
-
-# ORM - typeorm/sequelize
