+++ package.json
@@ -18,11 +18,14 @@
   },
   "homepage": "https://github.com/dr-shakya/itonics-node-training#readme",
   "dependencies": {
-    "express": "^4.17.3"
+    "express": "^4.17.3",
+    "pg": "^8.7.3",
+    "typeorm": "^0.2.45"
   },
   "devDependencies": {
     "@types/express": "^4.17.13",
     "@types/node": "^17.0.21",
+    "dotenv": "^16.0.0",
     "nodemon": "^2.0.15",
     "ts-node": "^10.6.0",
     "typescript": "^4.6.2"
