+++ src/Contract/item.ts
@@ -1,4 +1,4 @@
-export interface Items {
+export interface Item {
   id?: number;
   title: string;
   body: string;
