+++ src/Contract/student.ts
@@ -1,0 +1,4 @@
+export interface Student {
+  name: string;
+  address: string;
+}
