+++ src/Router/students.router.ts
@@ -1,10 +1,10 @@
 import express, { Request, Response } from 'express';
-import { students } from '../students';
+import { mockStudents } from '../Mock/mockData';
 
 const studentsRouter = express.Router();
 
 studentsRouter.get('', (req: Request, res: Response) => {
-  res.status(200).json(students);
+  res.status(200).json(mockStudents);
 });
 
 export default studentsRouter;
