--- src/students.ts
@@ -1,15 +1,0 @@
-export interface Student {
-  name: string;
-  address: string;
-}
-
-export const students: Array<Student> = [
-  {
-    name: 'Joe',
-    address: 'KTM'
-  },
-  {
-    name: 'John',
-    address: 'PKR'
-  }
-];
